requirements.txt
streamlit
yfinance
numpy
pandas
plotly
scipy
python-dotenv
google-generativeai
